+++ components/asic/include/bm1397.h
@@ -29,7 +29,7 @@ void BM1397_set_job_difficulty_mask(int);
 void BM1397_set_version_mask(uint32_t version_mask);
 int BM1397_set_max_baud(void);
 int BM1397_set_default_baud(void);
-void BM1397_send_hash_frequency(float frequency);
+void BM1397_send_hash_frequency(double frequency);
 task_result * BM1397_process_work();
 
 #endif /* BM1397_H_ */
